Remove subnavigation items from primary navigation in desktop and mobile
